Add validation to remap_users.py
This validation ensures that every user account that we find in the
users section of the json dump must appear in one of the user mapping
files. This is just a sanity check to ensure that users don't "sneak in"
under our noses at the last minute and helps with mapping validation.

Also, some minor improvements to the README to help users formulate
search queries for their JSON exports.