Always use email-matching username
The JIRA import does not respect explicitly over-ridden usernames if
it finds a username with the same email address that it
prefers. Multiple usernames can have the same email address. JIRA will
prefer one of them, but it is not clear how that preference is
set. THis patch tries to get the matching closer to the one JIRA
prefers by at least picking of the cnadidate usernames with the same
email address.